fix(android): make HMS ML Kit dependencies optional to resolve allowBackup conflict (closes #144)
